+++ lib/maru/types/json.ex
@@ -7,6 +7,7 @@ defmodule Maru.Types.Json do
 
   @doc false
   def parse(input, _) do
-    input |> Poison.decode!()
+    json_library = Maru.json_library()
+    json_library.decode!(input)
   end
 end